requirements.txt
fastapi~=0.116.0
pandas~=2.3.1
transformers~=4.53.2
huggingface-hub~=0.33.4
scikit-learn~=1.7.0
matplotlib~=3.10.3
scipy~=1.16.1
uvicorn~=0.35.0
langchain~=0.3.26
faiss-cpu~=1.11.0
nltk~=3.9.1
python-dotenv~=1.1.1
sentence-transformers~=5.0.0
langchain-community~=0.3.27
pydantic~=2.6.4


